float to double
git-svn-id: https://svn.salilab.org/imp/trunk@2106 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ kernel/src/Model.cpp
@@ -86,7 +86,7 @@ Float Model::evaluate(bool calc_derivs)
 
   // evaluate all of the active restraints to get score (and derivatives)
   // for current state of the model
-  Float score = 0.0;
+  double score = 0.0;
   DerivativeAccumulator accum;
   DerivativeAccumulator *accpt = (calc_derivs ? &accum : NULL);
